refactor: CookieManager utils 디렉토리로 이동
