+++ backend/src/main/java/kr/kro/colla/auth/presentation/AuthController.java
@@ -1,6 +1,9 @@
 package kr.kro.colla.auth.presentation;
 
+import kr.kro.colla.auth.domain.LoginUser;
+import kr.kro.colla.auth.presentation.argument_resolver.Authenticated;
 import kr.kro.colla.auth.service.AuthService;
+import kr.kro.colla.utils.CookieManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseCookie;
@@ -23,6 +26,11 @@ public class AuthController {
         return ResponseEntity.ok()
                 .header(HttpHeaders.SET_COOKIE, cookie.toString())
                 .build();
+    }
+
+    @GetMapping("/log")
+    public ResponseEntity Log(@Authenticated LoginUser loginUser) {
+        return ResponseEntity.ok(loginUser);
     }
 
 }
